Adds user auth for projects

+++ domain/aggregate.go
@@ -1,10 +1,12 @@
 package domain
 
 import (
+	"errors"
 	"time"
 
 	"github.com/MihaiBlebea/task-manager/domain/project"
 	"github.com/MihaiBlebea/task-manager/domain/task"
+	"github.com/MihaiBlebea/task-manager/domain/user"
 	"gorm.io/gorm"
 )
 
@@ -12,7 +14,14 @@ const (
 	timeLayout = "2006-01-02T15:04:05.000Z"
 )
 
+var (
+	ErrInvalidUserID = errors.New("Invalid user ID")
+	ErrUserNotOwner  = errors.New("User is not the owner")
+)
+
 type TaskManager interface {
+	RegisterUser(username, email, password string) (int, string, error)
+	LoginUser(email, password string) (int, string, error)
 	GetProject(userID, projectID int) (*project.Project, error)
 	GetUserProjects(userID int) ([]project.Project, error)
 	CreateProject(userID int, title, color, description, icon string) (int, error)
@@ -34,39 +43,90 @@ type TaskManager interface {
 }
 
 type taskManager struct {
+	userRepo    user.Repo
 	projectRepo project.Repo
 	taskRepo    task.Repo
 }
 
 func New(conn *gorm.DB) TaskManager {
 	return &taskManager{
+		userRepo:    user.NewRepo(conn),
 		projectRepo: project.NewRepo(conn),
 		taskRepo:    task.NewRepo(conn),
 	}
 }
 
-func (tm *taskManager) GetProject(userID, projectID int) (*project.Project, error) {
-	project, err := tm.projectRepo.FindWithID(projectID)
+func (tm *taskManager) RegisterUser(username, email, password string) (int, string, error) {
+	user, err := user.New(username, email, password)
 	if err != nil {
-		return project, err
+		return 0, "", err
+	}
+
+	id, err := tm.userRepo.Save(user)
+	if err != nil {
+		return 0, "", err
+	}
+
+	_, err = user.GenerateJWT()
+	if err != nil {
+		return 0, "", err
+	}
+
+	if err := tm.userRepo.Update(user); err != nil {
+		return 0, "", err
+	}
+
+	return id, user.Token, nil
+}
+
+func (tm *taskManager) LoginUser(email, password string) (int, string, error) {
+	u, err := tm.userRepo.FindWithEmail(email)
+	if err != nil {
+		return 0, "", err
+	}
+
+	if u.CheckPasswordHash(password) == false {
+		return 0, "", errors.New("Could not auth user")
+	}
+
+	return u.ID, u.Token, nil
+}
+
+func (tm *taskManager) GetProject(userID, projectID int) (*project.Project, error) {
+	if isValid := tm.validateUserID(userID); isValid == false {
+		return &project.Project{}, ErrInvalidUserID
+	}
+
+	proj, err := tm.projectRepo.FindWithID(projectID)
+	if err != nil {
+		return &project.Project{}, err
+	}
+
+	// Validate if the user is the owner of the project
+	if proj.ID != userID {
+		return &project.Project{}, ErrUserNotOwner
 	}
 
 	// Fetch all the tasks associated with this project
-	tasks, err := tm.taskRepo.FindTasksWithProjectID(project.ID)
+	tasks, err := tm.taskRepo.FindTasksWithProjectID(proj.ID)
 	if err != nil {
-		return project, err
+		return proj, err
 	}
 
 	if len(tasks) == 0 {
-		return project, nil
+		return proj, nil
 	}
 
-	project.Tasks = tasks
+	proj.Tasks = tasks
 
-	return project, nil
+	return proj, nil
 }
 
-func (tm *taskManager) GetUserProjects(userID int) ([]project.Project, error) {
+func (tm *taskManager) GetUserProjects(userID int) (projects []project.Project, _ error) {
+	if isValid := tm.validateUserID(userID); isValid == false {
+		return projects, ErrInvalidUserID
+	}
+
 	return tm.projectRepo.FindWithUserID(userID)
 }
 
@@ -76,12 +136,27 @@ func (tm *taskManager) CreateProject(
 	color,
 	description,
 	icon string) (int, error) {
+	if isValid := tm.validateUserID(userID); isValid == false {
+		return 0, ErrInvalidUserID
+	}
+
 	proj := project.New(userID, title)
+	proj.Color = color
+	proj.Description = description
+	proj.Icon = icon
 
 	return tm.projectRepo.Save(proj)
 }
 
 func (tm *taskManager) DeleteProject(userID, projectID int) error {
+	if isValid := tm.validateUserID(userID); isValid == false {
+		return ErrInvalidUserID
+	}
+
+	if isOwner := tm.validateProjectOwner(userID, projectID); isOwner == false {
+		return ErrUserNotOwner
+	}
+
 	return tm.projectRepo.Delete(projectID)
 }
 
@@ -92,6 +167,15 @@ func (tm *taskManager) UpdateProject(
 	color,
 	description,
 	icon string) error {
+
+	if isValid := tm.validateUserID(userID); isValid == false {
+		return ErrInvalidUserID
+	}
+
+	if isOwner := tm.validateProjectOwner(userID, projectID); isOwner == false {
+		return ErrUserNotOwner
+	}
+
 	proj := project.Project{
 		ID:          projectID,
 		UserID:      userID,
